Feature/parser system

+++ app_bot/broadcaster.py
@@ -1,12 +1,13 @@
 import asyncio
 import logging
+import pytz
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.triggers.cron import CronTrigger
-from datetime import datetime
+from datetime import datetime, timedelta
 from aiogram import Bot, types
 from aiogram.utils.i18n import I18n
 from core.database import init
-from core.database.models import User, Dispatcher, Post
+from core.database.models import User, Dispatcher, Post, CitiesForParser
 from mail_parser.mail_parser import mail_parser
 from settings import settings
 
@@ -90,23 +91,43 @@ class Broadcaster(object):
     # send mailing from admin panel
     @classmethod
     async def order_work(cls, order: Dispatcher):
-        try:
-            post = await Post.filter(id=(await order.post).id).first()
-        except Exception as e:
-            logger.error(f'Get post error', exc_info=e)
-            return
+        city: CitiesForParser | None = await order.city
+        if city:
+            # sending - check is it for city => call mail_parser
+            await mail_parser(bot=bot, city=city)
 
-        # sending
-        await cls.send_content_to_users(bot=bot, broadcaster_post=post, museum_id=order.museum_id)
+            # update city and order (+24 hours)
+            try:
+                city.was_sent = True
+                await city.save()
 
-        # delete order
-        try:
-            await Dispatcher.filter(id=order.id).delete()
-        except Exception as e:
-            logger.critical(f'Delete order error', exc_info=e)
-            return
+                await Dispatcher.filter(id=order.id).update(
+                    send_at=(datetime.now(pytz.timezone('Europe/Moscow')) + timedelta(days=1)).replace(
+                        hour=city.hour, minute=city.minute, second=0, microsecond=0
+                    ),
+                )
+            except Exception as e:
+                logger.critical(f'Update order for city {city.name} error', exc_info=e)
+                return
 
-        logger.info(f'order_id={order.id} has been sent to users')
+        # default mailing
+        else:
+            try:
+                post = await Post.filter(id=(await order.post).id).first()
+            except Exception as e:
+                logger.error(f'Get post error', exc_info=e)
+                return
+
+            await cls.send_content_to_users(bot=bot, broadcaster_post=post, museum_id=order.museum_id)
+
+            # delete order
+            try:
+                await Dispatcher.filter(id=order.id).delete()
+            except Exception as e:
+                logger.critical(f'Delete order error', exc_info=e)
+                return
+
+            logger.info(f'order_id={order.id} has been sent to users')
 
 
     @classmethod
@@ -123,6 +144,11 @@ class Broadcaster(object):
     async def start_event_loop(cls):
         logger.info('Broadcaster started')
         while True:
+            try:
+                await cls.__create_orders_for_mails()
+            except Exception as e:
+                logger.error(f'check_and_send_mails error', exc_info=e)
+
             try:
                 active_orders = await Dispatcher.filter(send_at__lte=datetime.now()).all()
                 logger.info(f'active_orders: {active_orders}')
@@ -163,6 +189,43 @@ class Broadcaster(object):
             await asyncio.sleep(settings.broadcaster_sleep)
 
 
+    @classmethod
+    async def __create_orders_for_mails(cls):
+        cities = await CitiesForParser.all()
+        if not cities:
+            return
+
+        tz = pytz.timezone('Europe/Moscow')
+        current_date = datetime.now(tz)
+
+        for city in cities:
+            # delete all active orders for cities if is_turn=False
+            if not city.is_turn:
+                await Dispatcher.filter(city_id=city.id).delete()
+                continue
+
+            # delete all active orders if city configuration was edited 20 seconds ago
+            if city.updated_at >= current_date - timedelta(seconds=10) and not city.was_sent:
+                await Dispatcher.filter(city_id=city.id).delete()
+
+            # check if there are already any orders for notifications, exit if there are any
+            orders_for_cities = await Dispatcher.filter(city_id=city.id)
+            if orders_for_cities:
+                continue
+            else:
+                # we schedule only for future time
+                send_at = current_date.replace(hour=city.hour, minute=city.minute, second=0, microsecond=0)
+                if current_date > send_at:
+                    send_at += timedelta(days=1)
+
+
+                # create order for city
+                await Dispatcher.create(
+                    city_id=city.id,
+                    send_at=send_at,
+                )
+
+
 async def main():
     await init()
     await Broadcaster.start_event_loop()
@@ -172,31 +235,6 @@ async def run_scheduler():
     scheduler = AsyncIOScheduler()
     scheduler.add_job(Broadcaster.send_notification,
                       trigger=CronTrigger(hour=settings.notification_hours, minute=settings.notification_minutes))
-
-    # add mail jobs
-    # nsk_krsk
-    scheduler.add_job(
-        func=mail_parser,
-        args=(bot, 'nsk_krsk', 17),
-        trigger=CronTrigger(hour=17, minute=0),
-        misfire_grace_time=10,
-    )
-
-    # samara
-    scheduler.add_job(
-        func=mail_parser,
-        args=(bot, 'samara', 20),
-        trigger=CronTrigger(hour=20, minute=0),
-        misfire_grace_time=10,
-    )
-
-    # nvg_spb
-    scheduler.add_job(
-        func=mail_parser,
-        args=(bot, 'nvg_spb', 21),
-        trigger=CronTrigger(hour=21, minute=0),
-        misfire_grace_time=10,
-    )
 
     scheduler.start()
